Fix performance issue in CalendarTemplate uses. Fixed some bugs in Calendar with leap years and size Added basic unit test for Calendar and CalendarTemplate

+++ src/31_calendar/CalendarTemplateElementTableSync.h
@@ -81,7 +81,7 @@ namespace synthese
 
 			static void Shift(
 				util::RegistryKeyType calendarId,
-				int rank,
+				size_t rank,
 				int delta,
 				boost::optional<db::DBTransaction&> transaction = boost::optional<db::DBTransaction&>()
 			);
